+++ lib/minimal_onboarding.dart
@@ -1,2 +1,3 @@
 export 'src/minimal_onboarding.dart';
 export 'src/onboarding_page_model.dart';
+export 'package:dots_indicator/src/dots_decorator.dart';