+++ lib/src/onboarding_page_model.dart
@@ -1,5 +1,7 @@
+/// Data model to create a single onboarding page
 class OnboardingPageModel {
   final String imagePath;
+  final String title;
   final String text;
-  OnboardingPageModel(this.imagePath, this.text);
+  OnboardingPageModel(this.imagePath, this.title, this.text);
 }