+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.0.0]
+
+- Add title in pages
+- Option to add custom decoration for dot progress indicator
+  
 ## [0.0.2]
 
 - Fix Onboarding App bar color issue